O_APPEND inserted on sys_open()
Insertion of the possibility of O_APPEND and some modifications on file.c
Now the offset on the openstruct structure is off_t type

+++ kern/include/file.h
@@ -53,9 +53,10 @@ struct openfile {
 
     mode_t mode; /*Read, Write, Read-Write ecc...*/
 
-    unsigned int offset;
+    off_t offset;
 
     unsigned int reference_count;
+
     struct lock *lock;   //By using a lock, multiple processes can use the openfile structure one at time, without conflicts.
 
 };
